core: grandpa: move justification to its own submodule
